update give helpful points
